hfspace/requirements.txt
gradio==4.44.0
libsql-client==0.5.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
aiofiles==23.2.1
chardet==5.2.0